+++ TDML/texture.cpp
@@ -18,7 +18,7 @@ namespace TDML
 
 	void texture::addNewId(int newID)
 	{
-		Log.output(newID);
+		Log.output((float)newID);
 		id.resize(ids+1);
 		id[ids] = newID;
 		ids++;
